+++ lib/Middleware/SubsonicMiddleware.php
@@ -16,6 +16,7 @@ use OCA\Music\AppFramework\BusinessLayer\BusinessLayerException;
 use OCA\Music\AppFramework\Core\Logger;
 use OCA\Music\Controller\SubsonicController;
 use OCA\Music\Db\AmpacheUserMapper;
+use OCA\Music\Service\LibraryFolderException;
 use OCA\Music\Utility\StringUtil;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\Response;
@@ -169,6 +170,8 @@ class SubsonicMiddleware extends Middleware {
 				);
 			} elseif ($exception instanceof BusinessLayerException) {
 				return $controller->subsonicErrorResponse(70, 'Entity not found');
+			} elseif ($exception instanceof LibraryFolderException) {
+				return $controller->subsonicErrorResponse(70, $exception->getMessage());
 			}
 		}
 		throw $exception;