+++ lib/Service/Scanner.php
@@ -501,7 +501,7 @@ class Scanner extends PublicEmitter {
 	private function getImageFiles(string $userId) : array {
 		try {
 			$folder = $this->librarySettings->getFolder($userId);
-		} catch (\OCP\Files\NotFoundException $e) {
+		} catch (LibraryFolderException $e) {
 			return [];
 		}